Finish Home and Footer
Only left to do is the rest of the components and then make it look good
on mobile.

The functionality was already done a couple of commits ago, these are
just modifications to make it look pretty although I might consider
modifying the router in the future so that the header doesn't get
re-render with each component.